+++ backends/cuda/CMakeLists.txt
@@ -339,6 +339,7 @@ install(
 # CUDA backend implementation
 set(_aoti_cuda_backend_sources runtime/cuda_backend.cpp
                                runtime/cuda_mutable_state.cpp
+                               runtime/cuda_weight_cache.cpp
 )
 if(_cuda_is_msvc_toolchain)
   # MSVC links aoti_cuda_backend into portable_lib without relying on C++
